fix(LicensesReport): proper fallback if no URL is defined for a package
Some npm packages in the wild, such as `reading-time`, do not define a
repository or homepage, or do not do so properly.

To avoid generating invalid anchor elements in such cases with no href
attribute, let's fall back to displaying a "-" text node in such cases,
as it's already done with the authors column.